Refactor dashboard tests
Removed usage of Floki
Why: Tying to HTML/CSS elements leads to fragile tests for no real
benefit. There was also a typo in the css class which would mean a
change in both the html page and the tests even though no business
behaviour changed. What you want to be testing is that the right data
appear on the page; the business logic is executed.

Used build_conn() instead of passing the map
Why: The map makes sense if you're passing other things with it as well
that are created on pre-test setup block. For a normal conn,
`build_conn()` reads better. Note: `conn()` will be deprecated hence the
usage of `build_conn()`

Removed extra `200` assertions
Why: All the tests exercise the same page so you only need 1 check for
the status which happens on the first test anw.